requirements.txt
PyQt5
selenium
python-dotenv
google-api-python-client
